oubuswebsite/oubusapi/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.0
Django==5.1
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
jwcrypto==1.5.6
mailjet-rest==1.3.4
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
